requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
django-environ==0.4.5
google-api-core==1.16.0
google-auth==1.10.1
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
setuptools==45.1.0
six==1.14.0
urllib3==1.25.7

